Внесение корректировок в процедуру работы с БД. Выност функций в отдельный файл.
