+++ index.php
@@ -1,8 +1,7 @@
 <?php
     require_once 'helpers.php';
-
+    require_once 'functions.php';
     date_default_timezone_set('Europe/Moscow');
-    $connectionDB = dbConnect();
 
     $postTypes = getPostTypes();
